Restructured test dir
It now contains a db and log dir that contains all the data needed to perform a test run at a 
later time.  Moved and renamed the existing clamav.sqlite3 db to *-OLD in the new test/db dir.
Updated config/clamav-test.yml to be able to repeatedly run using the db and logs in the test dir.