Cleaned up rails3 / 1.9 support
Fixed nasty empty rescue modifier issue with ruby 1.8.
Changed version of script to 1.1.0.
Added conditionals in default config/clamav.yml to handle Apple AppStore and direct download versions of ClamXav so that everything works out of the box instead of making people fiddle with the config in order to get it to run.